… Cargo.toml
… src/incrsan.rs
… src/internal_observer.rs
… src/kind/array_fold.rs
… src/kind/bind.rs
… src/kind/expert.rs
… src/kind/map.rs
… src/lib.rs
… src/node.rs
… src/node/id.rs
… src/node_update.rs
… src/public.rs
… src/scope.rs
… src/state.rs
… src/state/expert.rs
… src/var.rs
… tests/fixed_point.rs