Previously, pretty much any time you used an API on `Incr`, like `Incr::map`, rustc would basically have to compile the entire crate again. Because each closure was a new MapNode<F>, and each MapNode is a new NodeGenerics. This was the cause of pretty horrific compile times on any non-trivial project, including incremental's own test suite.

This was due to feeding through generic parameters all the way into Node, which was due to a misunderstanding on my part of how the ocaml compiler works. It checks types at compile time but nevertheless has runtime polymorphism. There is, for example, only one copy of `Node.recompute` in an OCaml binary. So rust generics are a bad way to emulate this.

Instead, this PR erases all generics except in the public APIs. All map functions are now boxed; all values are now boxed. There are trait objects everywhere. Runtime performance suffers, by between 10 and 70% on the microbenchmarks and negligible change to real-world perf. I imagine in bigger projects there will be a perf benefit to having only one copy of Node, which is MUCH friendlier to the instruction cache than *109 copies of Node and all its methods* etc. (Aside from  being able to compile your project in under 30 minutes, which is a perf win in its own way.)

This PR has no public API changes.

### Performance

If you do a lot of `Incr<f32>` and other types that are `usize` or smaller, then you will benefit from the `nightly-miny` feature. This swaps out std's Box for [`miny::Miny`](https://github.com/1e1001/miny). That crate is nightly-only, hence the feature name. It is between 10 and 30% faster than stock `incremental` on the `linear` benchmarks. In fact it is faster than before this PR on most of them. But unless all your map nodes are integer additions like in those microbenchmarks, then you're probably fine.

### Compile times

#### Before

```sh
; touch tests/basic.rs
; CARGO_INCREMENTAL=false cargo build --test basic
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 10.23s
; CARGO_INCREMENTAL=false cargo build --test basic --release
    Finished `release` profile [optimized] target(s) in 48.56s
```

#### After

```sh
; touch tests/basic.rs
; CARGO_INCREMENTAL=false cargo build --test basic
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.84s
; CARGO_INCREMENTAL=false cargo build --test basic --release
    Finished `release` profile [optimized] target(s) in 5.21s
```