@cormacrelf 2025-01-07
Draft because I wasn't paying attention to incrsan while doing all the refactoring, so `cargo check --features nightly-incrsan` fails with seven hundred and ninety-one errors.